+++ src/Facades/Hub.php
@@ -9,6 +9,7 @@ use Emeq\HubSdk\Http\HubConnector;
 use Emeq\HubSdk\Resources\Accounting;
 use Emeq\HubSdk\Resources\Accounts;
 use Emeq\HubSdk\Resources\Connections;
+use Emeq\HubSdk\Resources\ConnectSessions;
 use Emeq\HubSdk\Resources\Integrations;
 use Emeq\HubSdk\Resources\OAuth;
 use Illuminate\Support\Facades\Facade;
@@ -18,6 +19,7 @@ use Illuminate\Support\Facades\Facade;
  * @method static Accounts accounts()
  * @method static Integrations integrations()
  * @method static OAuth oauth()
+ * @method static ConnectSessions connectSessions()
  * @method static Connections connections()
  * @method static Accounting accounting()
  *